chore(core) Improve Input Group docs
