+++ apps/docs/src/app/core/component-docs/input-group/examples/index.ts
@@ -1,0 +1,25 @@
+import { InputGroupFormExampleComponent } from './input-group-form-example/input-group-form-example.component'
+import { InputGroupSearchExampleComponent } from './input-group-search-example/input-group-search-example.component';
+import {
+    InputGroupButtonExampleComponent,
+    InputGroupComplexExampleComponent,
+    InputGroupIconExampleComponent,
+    InputGroupStatesExampleComponent,
+    InputGroupTextCompactExampleComponent,
+    InputGroupTextExampleComponent
+} from './input-group-examples.component';
+
+export * from './input-group-search-example/input-group-search-example.component';
+export * from './input-group-form-example/input-group-form-example.component';
+export * from './input-group-examples.component';
+
+export const examples = [
+    InputGroupIconExampleComponent,
+    InputGroupTextExampleComponent,
+    InputGroupFormExampleComponent,
+    InputGroupButtonExampleComponent,
+    InputGroupStatesExampleComponent,
+    InputGroupSearchExampleComponent,
+    InputGroupComplexExampleComponent,
+    InputGroupTextCompactExampleComponent
+];
