+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-complex-example.component.html
@@ -1,21 +1,19 @@
-<fd-form-group>
+<div fd-form-item>
     <label fd-form-label>Input Group with complex button addon.</label>
     <fd-input-group>
-        <input fd-input-group-input fd-form-control />
-        <span fd-input-group-addon [button]="true">
-            <button fd-button [glyph]="'calendar'" [fdType]="'transparent'"></button>
+        <input fd-input-group-input fd-form-control/>
+        <span fd-input-group-addon button="true">
+            <button fd-button glyph="calendar" fdType="transparent"></button>
         </span>
     </fd-input-group>
-</fd-form-group>
-<br />
-<fd-form-group>
-    <div fd-form-item>
-        <label fd-form-label>Input Group with complex span addon.</label>
-        <fd-input-group>
-            <input fd-input-group-input fd-form-control />
-            <span fd-input-group-addon>
-                Addon
-            </span>
-        </fd-input-group>
-    </div>
-</fd-form-group>
+</div>
+
+<br/>
+
+<div fd-form-item>
+    <label fd-form-label>Input Group with complex span addon.</label>
+    <fd-input-group>
+        <input fd-input-group-input fd-form-control/>
+        <span fd-input-group-addon>Addon</span>
+    </fd-input-group>
+</div>
