+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-examples.component.ts
@@ -29,3 +29,9 @@ export class InputGroupTextCompactExampleComponent {}
     templateUrl: './input-group-complex-example.component.html'
 })
 export class InputGroupComplexExampleComponent {}
+
+@Component({
+    selector: 'fd-input-group-states-example',
+    templateUrl: './input-group-states-example.component.html'
+})
+export class InputGroupStatesExampleComponent {}
