--- apps/docs/src/app/core/component-docs/input-group/examples/input-group-form-example.component.html
@@ -1,10 +1,0 @@
-<form [formGroup]="customForm">
-    <fieldset fd-fieldset>
-        <div fd-form-item>
-            <label fd-form-label>Search Input</label>
-            <fd-input-group formControlName="searchQuery" [glyph]="'decline'" [button]="true" [placeholder]="'Search'">
-            </fd-input-group>
-            Search Query Entered: {{ customForm.controls.searchQuery.value }}
-        </div>
-    </fieldset>
-</form>
