+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-form-example/input-group-form-example.component.html
@@ -1,0 +1,13 @@
+<form [formGroup]="customForm">
+    <fieldset fd-fieldset>
+        <div fd-form-item>
+            <label fd-form-label>Search Input</label>
+            <fd-input-group glyph="decline"
+                            button="true"
+                            placeholder="Search"
+                            formControlName="searchQuery">
+            </fd-input-group>
+        </div>
+        <small>Search Query Entered: {{ customForm.controls.searchQuery.value }}</small>
+    </fieldset>
+</form>
