+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-icon-example.component.html
@@ -1,5 +1,17 @@
-<label fd-form-label>Left Aligned Icon Add-on</label>
-<fd-input-group [placement]="'before'" [glyph]="'accelerated'" [placeholder]="'Amount'"> </fd-input-group>
-<br />
-<label fd-form-label>Right Aligned Icon Add-on</label>
-<fd-input-group [placement]="'after'" [glyph]="'accelerated'" [placeholder]="'Amount'"> </fd-input-group>
+<div fd-form-item>
+    <label fd-form-label>Left Aligned Icon Add-on</label>
+    <fd-input-group placement="before"
+                    glyph="accelerated"
+                    placeholder="Amount">
+    </fd-input-group>
+</div>
+
+<br/>
+
+<div fd-form-item>
+    <label fd-form-label>Right Aligned Icon Add-on</label>
+    <fd-input-group placement="after"
+                    glyph="accelerated"
+                    placeholder="Amount">
+    </fd-input-group>
+</div>
