+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-states-example.component.html
@@ -2,65 +2,63 @@
     <label fd-form-label>Information</label>
     <fd-form-input-message-group>
         <fd-input-group
-            fd-form-control
-            [addOnText]="'Button'"
-            [button]="true"
-            [state]="'information'"
-            [placeholder]="'Placeholder'"
+            button="true"
+            addOnText="Button"
+            state="information"
+            placeholder="Placeholder"
         ></fd-input-group>
-        <fd-form-message [type]="'information'">
+        <fd-form-message type="information">
             Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
         </fd-form-message>
     </fd-form-input-message-group>
 </div>
 
-<br />
+<br/>
 
 <div fd-form-item>
     <label fd-form-label>Success</label>
     <fd-form-input-message-group>
         <fd-input-group
-            [addOnText]="'Button'"
-            [compact]="true"
-            [button]="true"
-            [state]="'success'"
-            [placeholder]="'Placeholder'"
+            button="true"
+            state="success"
+            addOnText="Button"
+            placeholder="Placeholder"
         ></fd-input-group>
-        <fd-form-message [type]="'success'">
+        <fd-form-message type="success">
             Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
         </fd-form-message>
     </fd-form-input-message-group>
 </div>
 
-<br />
+<br/>
 
 <div fd-form-item>
     <label fd-form-label>Warning</label>
     <fd-form-input-message-group>
         <fd-input-group
-            [addOnText]="'Button'"
-            [button]="true"
-            [state]="'warning'"
-            [placeholder]="'Placeholder'"
+            button="true"
+            state="warning"
+            addOnText="Button"
+            placeholder="Placeholder"
         ></fd-input-group>
-        <fd-form-message [type]="'warning'">
+        <fd-form-message type="warning">
             Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
         </fd-form-message>
     </fd-form-input-message-group>
 </div>
 
-<br />
+<br/>
 
 <div fd-form-item>
     <label fd-form-label>Error</label>
     <fd-form-input-message-group>
         <fd-input-group
-            [addOnText]="'Button'"
-            [button]="true"
-            [state]="'error'"
-            [placeholder]="'Placeholder'"
+            button="true"
+            state="error"
+            addOnText="Button"
+            placeholder="Placeholder"
         ></fd-input-group>
-        <fd-form-message [type]="'error'">
+        <fd-form-message type="error">
             Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
         </fd-form-message>
     </fd-form-input-message-group>
