--- apps/docs/src/app/core/component-docs/input-group/examples/input-group-states-example/input-group-states-example.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-input-group-states-example',
-    templateUrl: './input-group-states-example.component.html',
-    styleUrls: ['./input-group-states-example.component.scss']
-})
-export class InputGroupStatesExampleComponent {}
