+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-text-compact-example.component.html
@@ -1,6 +1,20 @@
-<label fd-form-label>Compact size</label>
-<fd-input-group [placement]="'before'" [addOnText]="'$'" [placeholder]="'Amount'" [compact]="true"> </fd-input-group>
+<div fd-form-item>
+    <label fd-form-label>Compact size</label>
+    <fd-input-group addOnText="$"
+                    compact="true"
+                    placement="before"
+                    placeholder="Amount">
+    </fd-input-group>
+</div>
 
-<label fd-form-label>Compact size button</label>
-<fd-input-group [placement]="'after'" [button]="true" [glyph]="'cart'" [placeholder]="'Amount'" [compact]="true">
-</fd-input-group>
+<br/>
+
+<div fd-form-item>
+    <label fd-form-label>Compact size button</label>
+    <fd-input-group glyph="cart"
+                    button="true"
+                    compact="true"
+                    placement="after"
+                    placeholder="Amount">
+    </fd-input-group>
+</div>
