+++ apps/docs/src/app/core/component-docs/input-group/input-group-docs.component.html
@@ -1,4 +1,4 @@
-<fd-docs-section-title [id]="'text-add'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="text-addon" componentName="inputGroup'">
     Text add-on
 </fd-docs-section-title>
 
@@ -13,7 +13,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'icon-addOn'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="icon-addon" componentName="inputGroup">
     Icon add-on
 </fd-docs-section-title>
 <description>
@@ -26,7 +26,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'button-addOn'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="button-addon" componentName="inputGroup">
     Button add-on
 </fd-docs-section-title>
 <description>
@@ -40,7 +40,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'compact'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="compact" componentName="inputGroup">
     Compact mode
 </fd-docs-section-title>
 <description> Adding <code>[compact]="true"</code> to the component makes the component compact. </description>
@@ -51,7 +51,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'inputGroupSearch'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="inputGroupSearch" componentName="inputGroup">
     Input Group Search
 </fd-docs-section-title>
 <description>
@@ -65,7 +65,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'reactiveForms'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="reactiveForms" componentName="inputGroup">
     Input Group Search Component Within Angular Reactive Forms
 </fd-docs-section-title>
 <description>It is possible to use the input-group-search component within an Angular Reactive Form.</description>
@@ -76,7 +76,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'complex'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="complex" componentName="inputGroup">
     Input Group with complex templates.
 </fd-docs-section-title>
 <description
@@ -90,7 +90,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title [id]="'states'" [componentName]="'inputGroup'">
+<fd-docs-section-title id="states" componentName="inputGroup">
     Input Group with States.
 </fd-docs-section-title>
 <description> </description>
