+++ apps/docs/src/app/core/component-docs/input-group/input-group-header/input-group-header.component.html
@@ -1,8 +1,8 @@
 <header>Input Group</header>
-<description
-    >The Input groups component are form inputs with add-ons that allows the user to better understand the information
-    been entered.</description
->
+<description>
+    The Input groups component are form inputs with add-ons that allows the user to better understand the information
+    been entered.
+</description>
 <import module="InputGroupModule"></import>
 
 <fd-header-tabs></fd-header-tabs>
