+++ apps/docs/src/app/core/component-docs/input-group/input-group-header/input-group-header.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'app-input-group-header',
-    templateUrl: './input-group-header.component.html',
-    styleUrls: ['./input-group-header.component.scss']
+    templateUrl: './input-group-header.component.html'
 })
 export class InputGroupHeaderComponent {}
